Fix soundtouch warning
See upstream fix:
https://gitlab.com/soundtouch/soundtouch/-/merge_requests/7/

+++ Externals/soundtouch/BPMDetect.cpp
@@ -561,7 +561,7 @@ float BPMDetect::getBpm()
 /// \return number of beats in the arrays.
 int BPMDetect::getBeats(float *pos, float *values, int max_num)
 {
-    int num = beats.size();
+    int num = (int)beats.size();
     if ((!pos) || (!values)) return num;    // pos or values NULL, return just size
 
     for (int i = 0; (i < num) && (i < max_num); i++)
